How 2 use Variables in pug file
First we must use #{}. It means inside of it is variable.
I used "pageTitle". It alone can't be variable. 
For example, the home.pug extends base.pug. And then the base.pug is activated by Controller file called videoController.js. Inside of it there would be re.render(). The important thing is res.render() can have 2 args. One thing is which pug file are you gonna render, and the other one is the arguments(made like object) you are gonna send to that pug file. So I wrote pageTitle and potato. 
As you can see home.pug, variable "potato" and because it extends base.pug, pageTitle would be "Home".