Iteration
Iteration means "반복", "repeating". I just delete the fakeUser and added videos obj and send it to home.pug.
Let's look how it works. From line 5 in home.pug,
I wrote 
"each video in videos"
    "li=video"
Let's see video as X and videos as Y.
Y should be same as the name the variable sended by trending func in videoController.js. But, video which is X can be called as anything like 
each potato in videos
    li= potato
is okay.