Mixin
Mixin is a different version of partial folder.
Mixin can get some data and update the part of itself using the data. And then for example +video is a "+mixin's name" and the name of info is a potato which is item of videos. See the mixin/video.pug.
It looks like function. So it must be "included". There is a "include mixins/video" line2. And the argment "info" come to the function. Finally for instance, info.rating could be potato.rating and the potato.rating means the rating that item of videos array has.
+++In home.pug, see the else next to each potato in videos. Pug is smart, so pug can recognize if there are items or nothing. If nothing, we can use else and write some code. That code will work when there is no item in the array.