Installing MongoDB, Mongoose and made a "model"
See ./models/Video.js. MongoDB starts with "Schema".
It define the form of the data and that data will make a big box of data. That is Schema. Using this, We can make a database, not making a fake DB. Formerly, we used a array and when we do CRUD work(Create, Read, Update, Delete), we must change the array's componenets directly. By using model and schema, we doesn't need to do that boring thing.