Difference of callback and promise
Video.find({}, callback func). Callback func is a way of waiting something. Before finding the {}(It means all stuff in the Video), the callback func does't work.
See line 4+~9+. We can anticipate that the console will say "Starting Search", "Search Finished", "I should be the last one",  "GET / ~~~~". But the console will say "Starting Search", "I should be the last one",  "GET / ~~~~", "Search Finished". 
The reasons is database is outside of javascript. So, the velocity of handling the database is slower than javascript. The result is come from the delayed time of getting some data from db.