Adding something to schema
https://mongoosejs.com/docs/schematypes.html<-is the important script.
The reason why giving max and minlength to the both upload.pug and Video.js is the following.
It is to protect the rule by twice.
If someone else hacked the html element in upload.pug and the max and minlength was gone. Then, If we don't have max and minlength in the schema, people will send the wrong data that I don't intended. So double protection can be made.