Making getEdit and postEdit func (1) and error page "404"
error meassage is activated when some user give the id that db does't have. getEdit give the data to edit.pug and edit.pug will fill the form by using video obj. And there is a one thing to see. "hashtags" is a string array and they are seperated. So we should use join() in edit.pug line8+. Second, videoController line35+. In those surrounded code are changing the original obj into modified one. But 35+, it helps if there is "#", not to add more "#". If it doesn't exist, everytime you edit hashtags part, you must delete all of hashtags in each one. It means if there is a string starting with "#", just return word same or word that starting with "#"