Making getEdit and postEdit func (2)
When we post the video, we don't need to get video obj writing  Video.findById(). Because when we redirect it(get request), getEdit will find the video obj and show us the screen. So, we just need to update the data in db. Before we update, we can use exist() func. It need filter. For example filter can be {title:"Hello"}. It means if there is a obj that has a title called "Hello", exist() will return true if not ,false. And then, FindByIdAndUpdate let us update the obj inside of Video model instantly.