Middleware in Mongoose not in express
https://mongoosejs.com/docs/middleware.html#types-of-middleware
In this link there are lots of type of middleware in Mongoose. Middleware helps the database to do something before for example like before save or update middleware can help me to change some data. 
In videoController, I just send hashtags. So the sended one is just a array that has one item. It is very hard job to write hashtags:hashtags.split(",").map()blablablabla~~ everytime before I save the hashtags. So It is repetitive job before "save".
By using Middleware in Mongoose we can resolve.
See the /models/Video.js line18+. Middleware should be made before model is made. "this" means the video obj that i want to save. In "this" there is a same obj and hashtags. Split func works only to the string. So, the hashtags[0] is the string. Just "hashtags" with out index is a array because the schema defined hashtags as string array and everytime user type it, it will be string array.. Then everytime it is saved, similarly with event, before "save" the middleware will seperate the hashtags.