Static function in schema file.
Static function is a tool that can make us not repeat the same function. As the line3+~5+ saying we can export the function. But If that function got too many, we must import many many function's name. By using static function we can make the function like Video.formatHashtags. It is kind of function that come from Video!. So we don't need to import all the stuffs everytime we use the function. All you need to do is import only "Video" and write like Video.formatHashtags. It is like a Video.create() and Video.findByIdAndUpdate().