Form Validation
Making a error message if there is a same username or email in the DB. To make it  I just modify the userController. If the same username or email exist then join page with errorMessage will show up. Focus on the operator "$or:[{}]" in line7+. {username:req.body.username} can be written like {username}. And Password Confirmation was also added.