Login part (1)
To make Login page -> I just copy paste the join page and changed the a(href) part. And then, I made 
rootRouter.route("/login").get(getLogin).post(postLogin);
I just made a change in videoController to add some status code for error page. Error page need status code starting with 4__ for example 400. Some google chrome would try to remember the password and ID. "res.render" alone doesn't mean that "this page have some error". To express this, I just added status code "400". If there is no error and it is successful, express will send status code starting with 2__ and Redirection Message will say 3__.
In userController, Lots of things have been changed.
See the postJoin. postJoin will gonna check the information that user send if there is a username that other person is using and if the password confirmation is failed. And after theses checking, there would be not returned also the await User.create func will be activated.
In post Login, the information that user typed in the login form will be sended to postLogin. postLogin will check if there is a same username in the DB, and password is matching with. It was't completed so next will say more.