Sessions and Session Id, Cookie, and Making Login part.
Let's make our server that the user is logged in!
To make our server make sense, we must use "session". Session is a history that user come or not. When user is trying to access some url, if we installed "session", express(backend part) will give a "session ID" to browser and browser will save that session ID in to the "Cookie". Cookie has lot of letters like this "blablalbalbl123213" but, Cookie got session ID itself.
After express send that to browser, express will also have a object "Session" and it contains a "session ID". Everytime browser makes a request to the server. browser will send the cookie that server gave already to the server automatically. In that cookie there is session Id, and by checking that session Id server will recognize that user is connecting