Intalling babel-loader, webpack and load the javascript code to pug file.
When we write the javascrpt code for browser(not for server), we need to use webpack. Webpack is already installed for next.js, react.js for example. But if we use framework that doesn't have default, we can set up intentionally. We must make "webpack.config.js" file. It must be same file name all the time. The rest of information will be on the code.