MongoStore Installation
Go to express connect-mongo. MongoStore make the session information get into the MongoDB we use. So if we use MongoStore, backend will save the session and sessionID and other stuffs into MongoDB instead of server memory. Then, cause we didn't save the session in server memory, even if we restart the server, the webpage will remember if we login or not.