How to use Environment variables and why we use env var
Environment variables should be used when we have something that has to be secret. In the case I made, the MongoDB URL and the COOKIE SECRET text should be hide from other person. If other person see our COOKIE SECRET text, cause it indicate the server which gave the cookie to user, other person can abuse it and we can be in trouble. Similarly, the exposure of URL of DB can occur a big problem. That means the whoever want to see DB information can access to DB.