Handling  some errors and making edit-profile distinguish if there is same username or email
line51+ and line58+ was modified cause 
"Error: Cannot set headers after they are sent to the client" was sent to me. So return was added.
In addition, Line79+~92+ was added for distinguishing if there is same username or email while editing pro-file. Edit-profile form need to refuse the username or the email same with each one in the DB. We defined both of them unique.