Bug Fixing(protecting postEdit and deleteVideo)(preventing user's password from postUpload's user.save())
If other user write the same url that can edit or delete that video, it is not reasonable. Only the owner of video can edit or delete it. So, it must also be protected by backend. If the owner of video doesn't match with the loggedin user, then it will redirect to home("/").
Second,  Every time we upload the video, I will load the id of video to UserModel. Doing this, user.save() activate the "presave" function that hash user.password. it means everytime I upload video, the password of mine will be hashed again and I would not login again cause the password would have been changed. So using this.isModified("password"), only when user.password is modified, password will be hashed.
When we upload the video, we doesn't change user.password. So it will not hash it.